requirements.txt
# Core
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

# n8n API
httpx==0.25.2

# AI & Embeddings
openai==1.6.1
anthropic==0.8.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1

# Vector Databases (choose based on your preference)
pinecone-client==3.0.0
weaviate-client==3.26.0
qdrant-client==1.7.0
chromadb==0.4.22

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Utilities
tiktoken==0.5.2
numpy==1.26.2
pandas==2.1.4
